Create Bricks class
